Fix MSK connector issues (#917)
Normally Athena queries are of the following structure:
SELECT * FROM <catalog name>.<schema name>.<table name>

The previous version of this connector did not map this structure
properly onto the structure for querying schemas in the Glue Schema Registry.

With this patch, this is how we map to the Glue Schema Registry now:
SELECT * FROM <catalog name for msk>.<glue registry name>.<glue schema name>

And then inside the Glue Schema, the "topicName" corresponds to the
Kafka topic name.

This means that the registry ARN environment variable is no longer
necessary and has been removed, since the registry is now contained in
the query structure itself.

This means that we also had to properly implement `doListSchemaNames`,
which was just returning `default` before, when it should actually be
listing out the Glue registries. However, note that we will only list out
Glue registries that have been "marked" for this connector.
  - To mark a Glue registry for this connector, make sure the string:
    {AthenaFederationMSK} is present within the `Description` of the
    Glue registry that you want to mark.

This patch also removes a lot of extraneous wrapping, also adds
pagination support for various list* functions, and finally also adds
case insensitive lookups into the Glue Schema Registry.